+++ sdk/metric/meter_test.go
@@ -382,9 +382,9 @@ func TestMeterCreatesInstruments(t *testing.T) {
 			},
 			want: metricdata.Metrics{
 				Name: "histogram",
-				Data: metricdata.Histogram{
+				Data: metricdata.Histogram[int64]{
 					Temporality: metricdata.CumulativeTemporality,
-					DataPoints: []metricdata.HistogramDataPoint{
+					DataPoints: []metricdata.HistogramDataPoint[int64]{
 						{
 							Attributes:   attribute.Set{},
 							Count:        1,
@@ -446,9 +446,9 @@ func TestMeterCreatesInstruments(t *testing.T) {
 			},
 			want: metricdata.Metrics{
 				Name: "histogram",
-				Data: metricdata.Histogram{
+				Data: metricdata.Histogram[float64]{
 					Temporality: metricdata.CumulativeTemporality,
-					DataPoints: []metricdata.HistogramDataPoint{
+					DataPoints: []metricdata.HistogramDataPoint[float64]{
 						{
 							Attributes:   attribute.Set{},
 							Count:        1,
@@ -489,7 +489,7 @@ func TestRegisterNonSDKObserverErrors(t *testing.T) {
 	mp := NewMeterProvider(WithReader(rdr))
 	meter := mp.Meter("scope")
 
-	type obsrv struct{ instrument.Asynchronous }
+	type obsrv struct{ instrument.Observable }
 	o := obsrv{}
 
 	_, err := meter.RegisterCallback(
@@ -1124,8 +1124,8 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 			},
 			wantMetric: metricdata.Metrics{
 				Name: "sfhistogram",
-				Data: metricdata.Histogram{
-					DataPoints: []metricdata.HistogramDataPoint{
+				Data: metricdata.Histogram[float64]{
+					DataPoints: []metricdata.HistogramDataPoint[float64]{
 						{
 							Attributes:   attribute.NewSet(attribute.String("foo", "bar")),
 							Bounds:       []float64{0, 5, 10, 25, 50, 75, 100, 250, 500, 750, 1000, 2500, 5000, 7500, 10000},
@@ -1206,8 +1206,8 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 			},
 			wantMetric: metricdata.Metrics{
 				Name: "sihistogram",
-				Data: metricdata.Histogram{
-					DataPoints: []metricdata.HistogramDataPoint{
+				Data: metricdata.Histogram[int64]{
+					DataPoints: []metricdata.HistogramDataPoint[int64]{
 						{
 							Attributes:   attribute.NewSet(attribute.String("foo", "bar")),
 							Bounds:       []float64{0, 5, 10, 25, 50, 75, 100, 250, 500, 750, 1000, 2500, 5000, 7500, 10000},
@@ -1254,7 +1254,7 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 	}
 }
 
-func TestAsynchronousExample(t *testing.T) {
+func TestObservableExample(t *testing.T) {
 	// This example can be found:
 	// https://github.com/open-telemetry/opentelemetry-specification/blob/8b91585e6175dd52b51e1d60bea105041225e35d/specification/metrics/supplementary-guidelines.md#asynchronous-example
 	var (
@@ -1277,7 +1277,7 @@ func TestAsynchronousExample(t *testing.T) {
 		const (
 			instName       = "pageFaults"
 			filteredStream = "filteredPageFaults"
-			scopeName      = "AsynchronousExample"
+			scopeName      = "ObservableExample"
 		)
 
 		selector := func(InstrumentKind) metricdata.Temporality { return temp }